+++ package.json
@@ -32,15 +32,12 @@
     "history": "^5.1.0",
     "html-react-parser": "^1.4.0",
     "lodash": "^4.17.21",
-    "quill": "^1.3.6",
     "react": "^17.0.2",
     "react-ckeditor-component": "^1.1.0",
     "react-dom": "^17.0.2",
     "react-html-parser": "^2.0.2",
     "react-icons": "^4.3.1",
     "react-intersection-observer": "^8.32.2",
-    "react-loading-skeleton": "^3.0.1",
-    "react-quill": "^1.3.5",
     "react-redux": "^7.2.6",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^4.0.3",
