+++ src/Common/Anchor.js
@@ -1,11 +1,11 @@
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 
-export default function Anchor(props) {
+export const Anchor = (props) => {
   const { children, to, _onClick } = props;
 
   return <ElAnchor to={to}>{children}</ElAnchor>;
-}
+};
 
 const ElAnchor = styled(Link)`
   display: block;
