+++ src/Common/index.js
@@ -1,11 +1,11 @@
-import Grid from './Grid';
-import Text from './Text';
-import Image from './Image';
-import Input from './Input';
-import Button from './Button';
-import Icon from './Icon';
-import Responsive from './Responsive';
-import Anchor from './Anchor';
-import Shimmer from './Shimmer';
+import { Grid } from './Grid';
+import { Text } from './Text';
+import { Image } from './Image';
+import { Input } from './Input';
+import { Button } from './Button';
+import { Icon } from './Icon';
+import { Responsive } from './Responsive';
+import { Anchor } from './Anchor';
+import { Shimmer } from './Shimmer';
 
 export { Grid, Text, Image, Input, Button, Icon, Responsive, Anchor, Shimmer };
