+++ src/Components/Modal.js
@@ -1,13 +1,13 @@
-import MenuApi from 'lib/api';
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { style } from './ModalStyle';
+import MenuApi from 'lib/api';
+import styled from 'styled-components';
+import { Grid, Text, Button } from 'Common';
 
-const Modal = (props) => {
+export const Modal = (props) => {
   const {
     title,
     content,
-    modalLink,
     postId,
     mainRef,
     deleteComment,
@@ -20,7 +20,7 @@ const Modal = (props) => {
 
   const closeModal = (e) => {
     if (e.target.innerText === '확인') {
-      command === 'postDelete' && onDeleteDetail();
+      command === 'postDelete' && onDeletePost();
       command === 'commentDelete' && onDeleteComment();
       command === 'goToBack' && history.push('/');
       command === 'saveLocalStorage' && onToggleModal();
@@ -29,9 +29,9 @@ const Modal = (props) => {
     }
   };
 
-  const onDeleteDetail = async () => {
+  const onDeletePost = async () => {
     await MenuApi.deleteDetail(postId);
-    history.push(modalLink);
+    history.push('/');
     onToggleModal();
   };
 
@@ -53,34 +53,57 @@ const Modal = (props) => {
     <ModalContainer onClick={(e) => closeModal(e)}>
       <ModalContent>
         <ModalHeader>
-          <ModalTitle>{title}</ModalTitle>
-          <ModalDesc>{content}</ModalDesc>
+          <Text size="26px" bold="700" margin="0 0 16px">
+            {title}
+          </Text>
+          <Text size="16px" color="rgb(73, 80, 87)">
+            {content}
+          </Text>
         </ModalHeader>
         <ModalFooter>
           <Button
-            style={{
-              background: 'rgb(233, 236, 239)',
-              color: 'rgb(73, 80, 87)',
-            }}
-            onClick={(e) => closeModal(e)}
+            bg="rgb(233, 236, 239)"
+            color="rgb(73, 80, 87)"
+            _onClick={(e) => closeModal(e)}
           >
             취소
           </Button>
-          <Button onClick={(e) => closeModal(e)}>확인</Button>
+          <Button _onClick={(e) => closeModal(e)}>확인</Button>
         </ModalFooter>
       </ModalContent>
     </ModalContainer>
   );
 };
 
-export default Modal;
+const ModalContainer = styled.div`
+  position: fixed;
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.4);
+  z-index: 999;
+`;
 
-const {
-  ModalContainer,
-  ModalContent,
-  ModalHeader,
-  ModalTitle,
-  ModalDesc,
-  ModalFooter,
-  Button,
-} = style;
+const ModalContent = styled.div`
+  width: 400px;
+  height: 220px;
+  display: flex;
+  flex-direction: column;
+  padding: 40px 30px 30px;
+  box-shadow: rgb(0 0 0 / 9%) 0px 2px 12px 0px;
+  background-color: #fff;
+`;
+
+const ModalHeader = styled.div`
+  flex: 1;
+`;
+
+const ModalFooter = styled.div`
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+`;
+
+export default Modal;
