+++ src/Components/PostShare.js
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
 import parse from 'html-react-parser';
-import { style } from './PostShareStyle';
+import styled from 'styled-components';
 import { removeHTMLTagFromString } from 'lib/removeHTMLTag';
 
-const PostShare = ({ isFixedShare, detailData }) => {
+export const PostShare = ({ isFixedShare, detailData }) => {
   const [shareContainerStyle, setShareContainerStyle] = useState({
     position: 'relative',
   });
@@ -85,6 +85,47 @@ const PostShare = ({ isFixedShare, detailData }) => {
   );
 };
 
-export default PostShare;
+const ShareContainer = styled.div`
+  position: absolute;
+  left: -7rem;
+`;
 
-const { ShareContainer, ShareWrap, ShareBtn, ShareImg } = style;
+const ShareWrap = styled.div`
+  position: ${(props) => props.position};
+  top: ${(props) => props.top};
+  width: 3.5rem;
+  height: 3.5rem;
+  background: rgb(248, 249, 250);
+  border: 1px solid rgb(241, 243, 245);
+  border-radius: 50%;
+  padding: 0.5rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+const ShareBtn = styled.button`
+  height: 3rem;
+  width: 3rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: white;
+  border: 1px solid rgb(173, 181, 189);
+  border-radius: 1.5rem;
+  color: rgb(134, 142, 150);
+  cursor: pointer;
+  z-index: 5;
+
+  &:hover {
+    color: black;
+    border: 1px solid black;
+  }
+`;
+
+const ShareImg = styled.svg`
+  width: 25px;
+  height: 25px;
+`;
+
+export default PostShare;
