--- src/Components/Tag/Tag.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { style } from './TagStyle';
-
-const Tag = ({ tagContent }) => {
-  return <Tagitem>{tagContent}</Tagitem>;
-};
-
-export default Tag;
-
-const { Tagitem } = style;
