--- src/Components/Tag/TagStyle.js
@@ -1,24 +1,0 @@
-import styled from 'styled-components';
-
-const Tagitem = styled.li`
-  margin-bottom: 0.875rem;
-  background: rgb(241, 243, 245);
-  padding-left: 1rem;
-  padding-right: 1rem;
-  margin-right: 0.9rem;
-  margin-bottom: 0.9rem;
-  height: 2rem;
-  border-radius: 1rem;
-  display: inline-flex;
-  align-items: center;
-  margin-right: 0.875 rem;
-  color: rgb(12, 166, 120);
-  text-decoration: none;
-  font-weight: 500;
-  font-size: 1rem;
-  cursor: pointer;
-`;
-
-export const style = {
-  Tagitem,
-};
