+++ src/Components/index.js
@@ -1,0 +1,25 @@
+import Card from './Card';
+import { Editor } from './Editor';
+import { ImgUpload } from './ImgUpload';
+import { Modal } from './Modal';
+import { Navbar } from './Navbar';
+import { PostShare } from './PostShare';
+import { ScrollToTop } from './ScrollToTop';
+import { Tag } from './Tag';
+import { DetailSkeleton } from './Skeleton/DetailSkeleton';
+import { ListSkeleton } from './Skeleton/ListSkeleton';
+import { Header } from './Header/Header';
+
+export {
+  Card,
+  Editor,
+  ImgUpload,
+  Modal,
+  Navbar,
+  PostShare,
+  ScrollToTop,
+  Tag,
+  DetailSkeleton,
+  ListSkeleton,
+  Header,
+};
